[config-plugins][config-types] set default url returned by getUpdateUrl in app.json
